web/backend/requirements.txt
paddlepaddle==2.3.2
paddlenlp==2.4.1
regex
fastcore
scikit-image
flask
flask_cors